Improve label output and warnings [revdep skip]
Catch the undesired column label of a named list with new grep call to handle the ExtractChartData call for Regression outputs. Also, improved the warning when variables are discarded in scatterplot outputs when there are elements that don't occur in both X and Y. Now the warning states the elements that have been removed when created the data.frame.