DS-2733 (#4)
* Add ability to use Regression output in Scatterplot

Update PrepareData and its internal functions to handle Regression output
via the input.data.raw argument

* Allow Y coordinate regression input in scatterplot

Y coordinates is a nested list structure. Logic modified to handle
regression inputs for scatterplots in the Y coordinate input aswell

* Prevent warning for Regression input [revdep skip]

A regression output is a list with more than 10 elements. This can cause the checks in prepareForSpecificCharts to throw a warning. This is not necessary since the Regression output is converted to a single variable later with ExtractChatrData

* Improve label output and warnings [revdep skip]

Catch the undesired column label of a named list with new grep call to handle the ExtractChartData call for Regression outputs. Also, improved the warning when variables are discarded in scatterplot outputs when there are elements that don't occur in both X and Y. Now the warning states the elements that have been removed when created the data.frame.

* Add unit tests for regression X & Y inputs to scatterplot [revdep skip]

* Update remotes to include flipRegression

flipRegression required for new unit tests.

* Add ability to have multiple raw Y inputs that include Regression

Improved handling of coerceToDataFrame to remap Y input.data.raw elements to a common named vector with NAs if element doesnt have an equivalent. Added unit tests to cover this case as well

* Retain rowlabels when more than two Regression elements in Y [revdep skip]