DS-2820 Allow more than one Y input for regression in Scatter (#8)
* DS-83 Allow more than one Y input for regression in Scatter

Previously having more than one Y input for a scatterplot that has a
regression model used in the inputs for either the X and Y coordinates
would throw an informative error. This commit allows the possibility of
having multiple Y inputs and gives an informative error message if there
is incompatible inputs in the X and Y coordinates. It does this by
checking the names of the points and matching accordingly.

* DS-2820 Remove unneeded line [ci skip]

Previously the Y input attributes where the prefix 'table.' was stripped
from the name attributes. This has been removed for consistency with
regression outputs.

* DS-2820 Amend warning message [revdep skip]