DS-5360 Don't always remove QTable attributes in PrepareData
Subscripted QTables had their attributes removed if it is identified
that the QTable class isn't permitted (ALLOW.QTABLE.CLASS or
allowQTables is defined properly) and the table is known to be
subscripted. Then to preserve old behaviour the attributes were removed.
However, verbs::SelectFromTable keeps attributes when subscripting as it
is needed to print the table nicelyy. These tables should produce nice
visualizations too.